Player Pick-ups
The functionality for the player to pick up resources is now in principal, done. However, there's weirdness with the sprite rendering(The player sprite is dissapearing and being replaced by the resources sprite; it should be appearing over their head)
Furthermore, I've noticed some quite noticeable issues with perspective, may need more eyes on it as I'm not the best judge of these sort of things, especially regarding the size of player & angle of camera. 
Have dropped in a rudimentary, alternative 3D player as well to see if it's any better. Will hold off on decision until Monday